Move utility functions to separate file to allow for testing
